+++ Chromagnon/chromformat.py
@@ -283,13 +283,13 @@ class ChromagnonReader(object):
         elif len(somewaves) >= 1: # the reference wavelength was not found but some found
             self.holder.refwave = somewaves[0]
             from PriCommon import guiFuncs as G
-            message = 'The original reference wavelength %i of the initial guess was not found in the target %s' % (self.refwave, self.holder.fn)
+            message = 'The original reference wavelength %i was not found in the target %s' % (self.refwave, self.holder.img.fn)
             G.openMsg(msg=message, title='WARNING')
 
         else:
             #self.holder.parm = N.zeros((self.rdr.nt, self.rdr.nw, self.num_entry), self.dtype)
             from PriCommon import guiFuncs as G
-            message = 'No common wavelength with initial guess was found in %s and %s' % (os.path.basename(self.file), self.rdr.file)
+            message = 'No common wavelength was found in %s and %s' % (os.path.basename(self.file), self.rdr.file)
             G.openMsg(msg=message, title='WARNING')
             return